Fix bug where NPCs could not engage in battle
There was a bug causing an early exit of the fight when one of the
players was an NPC. The issue stemmed from the .is_connected() function
on the PlayerType enum. It returned false automatically for all NPCs,
which could be considered a bug.

is_connected() has been modified to return a ConnectionStatus. There are
four variants:
* None, for when the question doesn't make sense
* Npc, for when the question doesn't make sense because they are NPC
* Connected, for when the player is human and connected
* Disconnected, for when the player is human and disconnected

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>